fix: marshal IShellBrowser and close remaining Win32 string contracts
Dynamic nss items (including stock io.file.create) run on a detached STA
and were calling the UI-thread IShellBrowser directly. Interface pointers
must be marshaled between apartments; CoMarshalInterThreadInterfaceInStream
is the one-shot primitive already used for shell-extension capture.

  https://learn.microsoft.com/windows/win32/com/single-threaded-apartments
  https://learn.microsoft.com/windows/win32/api/combaseapi/nf-combaseapi-comarshalinterthreadinterfaceinstream
  https://learn.microsoft.com/windows/win32/api/combaseapi/nf-combaseapi-cogetinterfaceandreleasestream

LoadStringW truncates at cchBufferMax-1; cchBufferMax 0 is the length
query. GetModuleFileNameExW succeeds on truncation and returns nSize-1,
and a non-NULL hModule needs PROCESS_QUERY_INFORMATION | PROCESS_VM_READ.

  https://learn.microsoft.com/windows/win32/api/winuser/nf-winuser-loadstringw
  https://learn.microsoft.com/windows/win32/api/psapi/nf-psapi-getmodulefilenameexw

Also keep the IComPtr SHBindToObject double-Release removal (out-parameter
conversion already owns the pointer), Release a transferred WIC bitmap if
init() fails, and the uncommitted whole-runtime QA that this branch stopped
on.

x64/x86 tests 25025/0; ARM64 compiles; MSI validator ok on all three.

+++ src/dll/src/Include/TaskbarHitCache.h
@@ -40,7 +40,7 @@ namespace Nilesoft
 			// larger than this, so a bucket cannot straddle "on a button" and
 			// "empty area" in a way that matters.
 			static constexpr long BUCKET = 16;
-			static constexpr uint32_t TTL_MS = 3000;
+			static constexpr uint64_t TTL_MS = 3000;
 			static constexpr size_t CAPACITY = 64;
 
 			static long bucket_of(long v) noexcept
@@ -50,7 +50,7 @@ namespace Nilesoft
 				return (v >= 0 ? v : (v - BUCKET + 1)) / BUCKET;
 			}
 
-			std::optional<bool> lookup(HWND taskbar, POINT pt, uint32_t now) const
+			std::optional<bool> lookup(HWND taskbar, POINT pt, uint64_t now) const
 			{
 				if(!taskbar)
 					return std::nullopt;
@@ -71,7 +71,7 @@ namespace Nilesoft
 				return std::nullopt;
 			}
 
-			void store(HWND taskbar, POINT pt, bool allowed, uint32_t now)
+			void store(HWND taskbar, POINT pt, bool allowed, uint64_t now)
 			{
 				if(!taskbar)
 					return;
@@ -130,7 +130,7 @@ namespace Nilesoft
 				long bx;
 				long by;
 				bool allowed;
-				uint32_t tick;
+				uint64_t tick;
 			};
 
 			mutable std::mutex _mutex;
